% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCA.R
\name{plot.PCA}
\alias{plot.PCA}
\title{Plot PCA}
\usage{
\method{plot}{PCA}(b)
}
\arguments{
\item{b}{an object of class "PCA"}
}
\value{
a screeplot
}
\description{
Plot PCA
}
\details{
This function creates Creates a screeplot to visualize principal components
}
