% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prince.R
\name{prince}
\alias{prince}
\title{PCA Function}
\usage{
prince(X, cor = T)
}
\arguments{
\item{X}{a matrix of data}
}
\value{
object of class "PCA"
}
\description{
Runs Principal Components Analysis on a dataset
}
\details{
This function creates an object with unique class "PCA" which includes loadings, eigenvectors
}
