% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyPrince.R
\name{shinyPrince}
\alias{shinyPrince}
\title{Shiny Application for PCA}
\usage{
shinyPrince(X)
}
\arguments{
\item{X}{a matrix of data}
}
\value{
opens a shiny server
}
\description{
Runs shiny app that explores PCA on a given dataset
}
\details{
This shiny app allows users to manipulate data that's inputted into PCA function, obtaining immediate visual results in the form of a screeplot and loadings info.
}
